# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Orange Money",
        "description": "API pour le système de mobile money Orange Money",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Récupérer les informations de l'utilisateur connecté",
                "operationId": "getCurrentUser",
                "responses": {
                    "200": {
                        "description": "Informations utilisateur récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur récupéré avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Mettre à jour le profil utilisateur",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "current_pin": {
                                        "description": "PIN actuel requis pour changer le PIN",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "new_pin": {
                                        "description": "Nouveau PIN (4 chiffres)",
                                        "type": "string",
                                        "example": "5678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil mis à jour avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/suppliers/balance-request": {
            "post": {
                "tags": [
                    "Fournisseurs"
                ],
                "summary": "Demander l'achat de solde (fournisseur)",
                "operationId": "requestBalance",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "montant"
                                ],
                                "properties": {
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Demande créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Demande de solde créée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès réservé aux fournisseurs approuvés"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "uuid-user"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Dupont"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Jean"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "705334611"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean.dupont@example.com"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "client",
                            "commercant",
                            "fournisseur"
                        ],
                        "example": "client"
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "inactif",
                            "en_attente"
                        ],
                        "example": "actif"
                    },
                    "pin": {
                        "description": "Code PIN à 4 chiffres",
                        "type": "string",
                        "example": "1234"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}